requirements.txt
aiohttp==3.8.0
pycryptodome==3.11.0
jwt==1.3.1
